+++ tests/test_client.py
@@ -62,6 +62,16 @@ class TestGetwaveformClass:
         assert os.path.exists(ctable)
         os.remove(ctable)
 
+    def test_get_waveform_starttime_in_string(self, client):
+        data, ctable = client.get_waveform('0101', '2010-01-01T00:00', 9)
+
+        assert data == '0101_201001010000_9.cnt'
+        assert os.path.exists(data)
+        os.remove(data)
+        assert ctable == '0101_20100101.ch'
+        assert os.path.exists(ctable)
+        os.remove(ctable)
+
     def test_get_waveform_custom_name_1(self, client):
         starttime = datetime(2010, 1, 1, 0, 0)
         data, ctable = client.get_waveform('0101', starttime, 1,
@@ -168,6 +178,12 @@ class TestGetCatalogClass:
     def test_get_arrivaltime_use_datetime(self, client):
         startdate = datetime(2010, 1, 1)
         data = client.get_arrivaltime(startdate, 5)
+        assert data == 'measure_20100101_5.txt'
+        assert os.path.exists(data)
+        os.remove(data)
+
+    def test_get_arrivaltime_startdate_in_string(self, client):
+        data = client.get_arrivaltime('20100101', 5)
         assert data == 'measure_20100101_5.txt'
         assert os.path.exists(data)
         os.remove(data)